# openapi.json
{"openapi":"3.0.1","info":{"title":"Jay Boot API","description":"Jay Boot 项目接口文档","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"认证管理","description":"用户注册、登录、会话与密码管理"},{"name":"RBAC 管理","description":"角色、权限与用户角色分配"},{"name":"系统状态","description":"系统健康检查接口"},{"name":"Billing 管理","description":"套餐与订阅管理"},{"name":"租户管理","description":"当前租户查询与更新"}],"paths":{"/api/tenants/current/update":{"post":{"tags":["租户管理"],"summary":"更新当前租户","operationId":"updateCurrent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TenantUpdateCurrentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTenantCurrentResponse"}}}}},"security":[{"satoken":[]}]}},"/api/rbac/users/{id}/roles":{"post":{"tags":["RBAC 管理"],"summary":"用户分配角色","operationId":"assignUserRoles","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAssignRolesRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}},"security":[{"satoken":[]}]}},"/api/rbac/roles":{"get":{"tags":["RBAC 管理"],"summary":"查询角色列表","operationId":"roles","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListRoleResponse"}}}}},"security":[{"satoken":[]}]},"post":{"tags":["RBAC 管理"],"summary":"创建角色","operationId":"createRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseRoleResponse"}}}}},"security":[{"satoken":[]}]}},"/api/rbac/roles/{id}/permissions":{"post":{"tags":["RBAC 管理"],"summary":"角色分配权限","operationId":"assignPermissions","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleAssignPermissionsRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}},"security":[{"satoken":[]}]}},"/api/billing/subscriptions":{"post":{"tags":["Billing 管理"],"summary":"创建当前租户订阅","operationId":"createSubscription","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubscriptionCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseSubscriptionView"}}}}},"security":[{"satoken":[]}]}},"/api/billing/subscriptions/{id}/update":{"post":{"tags":["Billing 管理"],"summary":"更新订阅状态","operationId":"updateSubscription","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubscriptionUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseSubscriptionView"}}}}},"security":[{"satoken":[]}]}},"/api/auth/token/refresh":{"post":{"tags":["认证管理"],"summary":"刷新令牌","operationId":"refreshToken","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthTokenResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["认证管理"],"summary":"用户注册","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthTokenResponse"}}}}}}},"/api/auth/password/change":{"post":{"tags":["认证管理"],"summary":"修改密码","operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/auth/logout":{"post":{"tags":["认证管理"],"summary":"用户登出","operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/auth/login":{"post":{"tags":["认证管理"],"summary":"用户登录","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthTokenResponse"}}}}}}},"/api/tenants/current":{"get":{"tags":["租户管理"],"summary":"查询当前租户","operationId":"current","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTenantCurrentResponse"}}}}},"security":[{"satoken":[]}]}},"/api/system/ping":{"get":{"tags":["系统状态"],"summary":"系统存活检查","operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/rbac/permissions/check":{"get":{"tags":["RBAC 管理"],"summary":"检查当前用户权限","operationId":"checkPermission","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePermissionCheckResponse"}}}}},"security":[{"satoken":[]}]}},"/api/billing/plans":{"get":{"tags":["Billing 管理"],"summary":"查询套餐列表","operationId":"plans","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPlanView"}}}}},"security":[{"satoken":[]}]}},"/api/auth/session":{"get":{"tags":["认证管理"],"summary":"获取当前会话","operationId":"session","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthSessionResponse"}}}}}}}},"components":{"schemas":{"TenantUpdateCurrentRequest":{"required":["name"],"type":"object","properties":{"name":{"maxLength":64,"minLength":2,"type":"string","description":"工作区名称","example":"Acme Workspace"}},"description":"更新当前租户请求"},"ApiResponseTenantCurrentResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/TenantCurrentResponse"}}},"TenantCurrentResponse":{"type":"object","properties":{"tenantId":{"type":"integer","description":"租户 ID","format":"int64","example":2001},"name":{"type":"string","description":"工作区名称","example":"workspace-a1b2"},"ownerUserId":{"type":"integer","description":"租户拥有者用户 ID","format":"int64","example":1001},"planCode":{"type":"string","description":"套餐编码","example":"FREE"}},"description":"当前租户信息"},"UserAssignRolesRequest":{"required":["roleIds"],"type":"object","properties":{"roleIds":{"type":"array","description":"角色 ID 列表","items":{"type":"integer","description":"角色 ID 列表","format":"int64"}}},"description":"用户角色分配请求"},"ApiResponseVoid":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"object"}}},"RoleCreateRequest":{"required":["name"],"type":"object","properties":{"name":{"maxLength":64,"minLength":2,"type":"string","description":"角色名称","example":"MANAGER"}},"description":"角色创建请求"},"ApiResponseRoleResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/RoleResponse"}}},"RoleResponse":{"type":"object","properties":{"roleId":{"type":"integer","description":"角色 ID","format":"int64","example":3001},"name":{"type":"string","description":"角色名称","example":"MANAGER"}},"description":"角色响应"},"RoleAssignPermissionsRequest":{"required":["permissionCodes"],"type":"object","properties":{"permissionCodes":{"type":"array","description":"权限编码列表","items":{"type":"string","description":"权限编码列表"}}},"description":"角色权限分配请求"},"SubscriptionCreateRequest":{"required":["planId"],"type":"object","properties":{"planId":{"type":"integer","description":"套餐 ID","format":"int64","example":1001},"initialStatus":{"type":"string","description":"初始状态，仅支持 TRIALING 或 ACTIVE，默认 TRIALING","example":"TRIALING"},"trialEndAt":{"type":"string","description":"试用结束时间","format":"date-time"},"currentPeriodEnd":{"type":"string","description":"当前计费周期结束时间","format":"date-time"}},"description":"创建订阅请求"},"ApiResponseSubscriptionView":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/SubscriptionView"}}},"SubscriptionView":{"type":"object","properties":{"subscriptionId":{"type":"integer","description":"订阅 ID","format":"int64","example":2001},"tenantId":{"type":"integer","description":"租户 ID","format":"int64","example":3001},"planId":{"type":"integer","description":"套餐 ID","format":"int64","example":1001},"status":{"type":"string","description":"订阅状态","example":"TRIALING"},"trialEndAt":{"type":"string","description":"试用结束时间","format":"date-time"},"currentPeriodEnd":{"type":"string","description":"当前计费周期结束时间","format":"date-time"},"effectiveTime":{"type":"string","description":"生效时间","format":"date-time"},"cancelTime":{"type":"string","description":"取消时间","format":"date-time"}},"description":"订阅信息"},"SubscriptionUpdateRequest":{"required":["targetStatus"],"type":"object","properties":{"targetStatus":{"type":"string","description":"目标状态","example":"ACTIVE"},"currentPeriodEnd":{"type":"string","description":"当前计费周期结束时间","format":"date-time"},"cancelTime":{"type":"string","description":"取消时间（目标状态为 CANCELED 时可传）","format":"date-time"}},"description":"更新订阅请求"},"ApiResponseAuthTokenResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AuthTokenResponse"}}},"AuthTokenResponse":{"type":"object","properties":{"token":{"type":"string","description":"访问令牌"},"tokenTimeout":{"type":"integer","description":"令牌有效期（秒）","format":"int64"},"user":{"$ref":"#/components/schemas/AuthUserView"}},"description":"认证令牌响应"},"AuthUserView":{"type":"object","properties":{"id":{"type":"integer","description":"用户 ID","format":"int64","example":1001},"tenantId":{"type":"integer","description":"租户 ID","format":"int64","example":2001},"email":{"type":"string","description":"邮箱","example":"user@example.com"},"status":{"type":"string","description":"用户状态","example":"ACTIVE"}},"description":"登录用户信息"},"RegisterRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","description":"邮箱","example":"user@example.com"},"password":{"maxLength":64,"minLength":8,"type":"string","description":"密码","example":"Password123"}},"description":"用户注册请求"},"ChangePasswordRequest":{"required":["newPassword","oldPassword"],"type":"object","properties":{"oldPassword":{"type":"string","description":"旧密码","example":"Password123"},"newPassword":{"maxLength":64,"minLength":8,"type":"string","description":"新密码","example":"Password456"}},"description":"修改密码请求"},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","description":"邮箱","example":"user@example.com"},"password":{"maxLength":64,"minLength":8,"type":"string","description":"密码","example":"Password123"}},"description":"用户登录请求"},"ApiResponseListRoleResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/RoleResponse"}}}},"ApiResponsePermissionCheckResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PermissionCheckResponse"}}},"PermissionCheckResponse":{"type":"object","properties":{"code":{"type":"string","description":"权限编码","example":"rbac.role.read"},"granted":{"type":"boolean","description":"是否已授予"}},"description":"权限检查响应"},"ApiResponseListPlanView":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PlanView"}}}},"PlanView":{"type":"object","properties":{"planId":{"type":"integer","description":"套餐 ID","format":"int64","example":1001},"code":{"type":"string","description":"套餐编码","example":"PRO"},"name":{"type":"string","description":"套餐名称","example":"专业版"},"billingCycle":{"type":"string","description":"计费周期","example":"MONTHLY"},"quotaJson":{"type":"string","description":"额度配置 JSON","example":"{\"users\":10,\"storageGb\":100}"},"price":{"type":"integer","description":"价格（最小货币单位）","format":"int64","example":19900},"status":{"type":"string","description":"套餐状态","example":"ACTIVE"}},"description":"套餐信息"},"ApiResponseAuthSessionResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AuthSessionResponse"}}},"AuthSessionResponse":{"type":"object","properties":{"loginId":{"type":"integer","description":"登录用户 ID","format":"int64"},"token":{"type":"string","description":"当前令牌"},"tokenTimeout":{"type":"integer","description":"令牌有效期（秒）","format":"int64"},"user":{"$ref":"#/components/schemas/AuthUserView"}},"description":"当前会话响应"}},"securitySchemes":{"satoken":{"type":"apiKey","name":"satoken","in":"header"}}}}